fix: request duration metric

+++ crates/runtime/src/util/metrics.rs
@@ -17,7 +17,7 @@ lazy_static! {
     static ref MEMORY_USAGE: HistogramVec = register_histogram_vec!("fastedge_wasm_memory_used", "WASM Memory usage", &["executor"]).unwrap();
 }
 
-pub fn metrics(result: AppResult, label: &[&str], duration: Option<u64>, memory_used: Option<u64>) {
+pub fn metrics(result: AppResult, label: &[&str], duration: Option<f64>, memory_used: Option<f64>) {
     TOTAL_COUNT.with_label_values(label).inc();
 
     if result != AppResult::SUCCESS {
@@ -34,9 +34,9 @@ pub fn metrics(result: AppResult, label: &[&str], duration: Option<u64>, memory_
     }
 
     if let Some(duration) = duration {
-        REQUEST_DURATION.with_label_values(label).observe(duration as f64);
+        REQUEST_DURATION.with_label_values(label).observe(duration);
     }
     if let Some(memory_used) = memory_used {
-        MEMORY_USAGE.with_label_values(label).observe(memory_used as f64);
+        MEMORY_USAGE.with_label_values(label).observe(memory_used);
     }
 }
